[add] animation to admin dashboard background
