+++ src/components/viewElection/components/ViewElection.css
@@ -11,7 +11,6 @@
 .electionItem{
     width: 450px;
     max-width: 700px;
-    /* min-width: 40px; */
     margin: 0px 10px 50px 10px;
 }
